WIP: Gradually refactoring task for addition of duplicate checking
